Avoid dependency cycle from apexer to apexer
Previously, the python host module apexer had itself in the required
property, forming a dependency cycle. So far, this has been harmless
because the installation of required modules is done by Make and Make is
not checking cycles there. However, this will prevent Soong from
implementing the installation of the required modules by itself.

Fixing the dependency problem.

Bug: 321626681
Test: builds
Change-Id: I13ff8ddbcca3099a71d18d5a91617b4bc6b3437e