Fix ApexCompressionTests
The test was broken because it used a different soong module for the
original apex. Because the NOTICE in .apex contains the intermediate
paths, a different soong module generate a different artifact even if
the content is same as other apex module.

To fix that, the test now uses the original apex from the same soong
module using "{.apex}" tag.

Bug: 338157001
Test: atest apex_compression_platform_tests:ApexCompressionTests
Change-Id: I618d31083de30511d8cec2a72c17c77bb12ac776