apexd-metrics: helper class to send atoms
This helper class sends "install requested" event (and possibliy "end"
event as well) on its dtor. To mark the success, Commit() should be
called. Otherwise, the installation request is considered as failure.

This is to clean up code and to prepare for the upcoming change to fill
"hal info". Since "hal info" can be obtained within "temp-mount", we'd
better delay sending events until temp-mounts.

Bug: 366217822
Test: atest ApexTestCases
Change-Id: Ib6bb78de2d9cd06204d9253aad799e87e976f3a3