Fix ApexdMountTest to clean up
ApexdMountTest uses its own copy of MountedApexDatabase() which is reset
in TearDown(). So, we can safely "deactivate" all apexes in TearDown()
to avoid accidental failures due to other test's failure.

Previously, this was done by explicit calls to UnmountOnTeardown() but
this doesn't work when a test fails (and leaves an apex as activated).

Bug: 383472293
Test: atest ApexTestCases
Change-Id: Iac8ab49cb975628bb49d750c4f2fc458afc438c9