apex-ls: lister for apex files
The new tool lists the contents of apex along with security
contexts (with -Z). This replaces `deapexer list`.

$ apex-ls -Z com.android.foo.apex

Note that the current deapexer doesn't support listing erofs apexes. It
would be ideal to add the support to deapexer but it's a bit difficult
to modify deapexer to have a new external dependency to dump.erofs for
this matter because it would break existing clients of the tool.

Instead, I'd like to keep the current deapexer to focus on "extract"
(which works with both ext4/erofs apexes), and in AOSP, we use this new
tool when listing apexes. The tool could be replaced with deapexer again
in the future when it supports listing erofs apexes.

Bug: 353896817
Test: OVERRIDE_PRODUCT_DEFAULT_APEX_PAYLOAD_TYPE=erofs m run_apex_sepolicy_tests
Change-Id: I2703dd69a3d509e7ba553843907a19f12ea40a4a