apexd: Force apex_payload.img to flush to disk on decompression.
Summary: We have noticed that valid .capex (compressed apex) files when
decompressed on certain chipsets appear to result in missing blocks in
the decompressed ext4 dm-verity apex_payload.img.

When apexd then goes to mount this decompressed ext4 image it fails on these corrupted blocks with a dm-verify error leading to a forced reboot.

By forcing the writes on the decompression path to flush to disk with O_SYNC we can rule out any odd edge cases around the page cache and mounting as closing the file has no guarantee of actually flushing the contents to disk.

Test: Ran roughly 100 times testing upgrades and decompressing apex files. It so far has not reproduced the crash.

Change-Id: Iccd7c5bffa7dee7bf9b9577ed197e269ac5f4a42
Signed-off-by: Abdelrahman Daim <adaim@meta.com>