Declare the apps included in prebuilt shim apexes
Make built apkcerts.txt currently contains an entry for every app in the
tree, but this will be changed to the list of installed apps for the
device. In preparation of this change, declare the apps included in the
prebuilt apexes to ensure that these can be signed with the correct cert
information.

Test: With RELEASE_APKCERTS_INSTALL_ONLY=true reapplied (https://r.android.com/3552561),
previously failed signing test now passes
https://android-build.corp.google.com/builds/abtd/run/L03300030010481704
Test: presubmits

Bug: 403528876
Bug: 399788149
Change-Id: I535ee8dc20cf8520f1f1fd9793a425d0a3f4af9a