Do not use ApexFileRepository::GetDataApex()
Data apexes stored in ApexFileRepository are a snapshot of apexes in
"active" state (apex files in /data/apex/active to be specific).
Just being in the active state doesn't guarantee that they are actually
being in used. (ApexFileRepository can have dangling references to
deleted apexes!)

MountedApexDatabase provides the correct list of "active" apexes.

Bug: 405014066
Test: ApexTestCases
Change-Id: I56349d93217ac2a7feb4397111d2caea9e9f5b54