Do not use loop.lo_crypt_name
Loop devices are created with lo_crypt_name set and it's checked before
destorying to avoid destroying unrelated loop devices.

However, lo_crypt_name is not supported:
https://lore.kernel.org/all/20211019075639.2333969-1-hch@lst.de/

It's been okay because loop devices are automaticallay cleared with
LO_FLAGS_AUTOCLEAR and the explicit LOOP_CLR_FD was never called.

If we call LOOP_CLR_FD anyway, it always fails with ENXIO because it's
cleared before the call.

Hence, removing DestroyLoopDevice as well.

Bug: 386296187
Test: atest ApexTestCases
Change-Id: I5f734e12e6584deefe4340afd34320ace2df1153