ApexImageManger manages active/backup list of images
With "mount-before-data", there's no /data/apex/active directory where
"active" apexes are found. Apexd needs to know the list of "active"
apexes. It also needs to update it when activating staged sessions.

The list should be available before the /data partition, hence it's
stored in the /metadata partition.

The reason for "backup" list is to recover the last-known-good list of
"active" when staged installation goes wrong. For the information stored
in the /metadata partition, we can simply rely on filesystem's
checkpoint support.

Bug: 404403537
Test: atest ApexTestCases
Change-Id: I6ff8b3a36e1fdf1e9ea10b2d89ed8029ce808363