Move RemoveInactiveDataApex() to BootCompletedCleanup()
RemoveInactiveDataApex() was in BootCompletedCleanup() before. It was
moved to OnStart() to fix the race between OnStart() and
InstallPackage().

Since we now have a mutex to avoid races, we can move it back to the
original location.

BootCompletedCleanup() is a more suitable place for cleanup because
OnStart() belongs to a hot path and we want to offload as much as we
can.

Bug: 365000206
Test: ApexTestCases
Change-Id: Ifa6d20b6f2e1396c62dc2f497d0ce03fe57087de