apexd starts with mount-before-data
Besides the flag guard, it needs additional condition: init has started
with a single mount namespace.

It's unfortunate that apexd relies on init for the decision that should
belong to apexd. But note that the apex activation needs a close
interaction between init and apexd, and that init starts before apexd.

Bug: 381173074
Test: m & device boots (with the flag on locally)
Change-Id: Ibb2a5d5a086738eb111bfb1ede6ffaaf17b9732d