apexd-bootstrap creates loop devices using hint
Getting a free loop device and configuring it (setting fd) need to be
done together. That's why the code is enclosed in a mutex guard.
However, this effectively serializes the apex activation threads. This
is not a big deal when loop devices are already created, but can be a
problem if not. With "mount-before-data", apexd-bootstrap will use many
loop devices. It spends most of its running time to wait for loop device
creation one-by-one. On the cuttlefish, apexd-bootstrap takes ~3sec,
which is not acceptable.

In this change, apex activation is done in parallel by specifying loop
device ids. Since it's the bootstrap stage, the loop ID ranges from 0 to
N-1 (N, the number of APEXes) should be fine to use. This way, loop
creation can be done in parallel without a mutex.

apexd-bootstrap's running time is reduced from ~3s to ~350ms.

Note that there's a rare case when some loop devices are already in use.
It's okay because there's a retry loop for loop creation and it falls
back to creating a new loop device instead of using an assigned ID.

Bug: 413173431
Flag: com.android.apex.flags.mount_before_data
Test: ApexTestCases
Change-Id: I6ed40a29bd3915d1e5bb33313e058780d24a3657