apexd: refactor around ActivateApexPackages()
Compressed apexes should be decompressed before activation. Now
ProcessCompressedApex() is called in ActivateApexPackages() to dedup and
a vector-variant of ProcessCompressedApex() overload is removed for
simplicity.

To hold the ApexFile objects created by ProcessCompressedApex(),
ActivationContext object is introduced and passed to Activate*
functions.

ActivationContext object in fact will be useful when we want to add
ThreadPool for activation, which should be passed around as well to
various functions.

Bug: 414244821
Test: ApexTestCases
Change-Id: I2c851bb001b733c249c07269ff6b228f25f8315f