apexd: creates mount_before_data onStart
/metadata/apex/config/mount_before_data is a marker file indicating the
feature is enabled. Since this file is created on markBootCompleted(),
the device which has started with the feature enabled can go back to
the feature-disabled state. This is very rare but can happen when the
device is rebooted before markBootCompleted() is called. It's not that
problematic because the device will enable the feature eventually when
markBootCompleted() is called.

This change reduces the possibility by creating the marker file as soon
as it starts with the feature enabled.

Bug: 421383225
Flag: com.android.apex.flags.mount_before_data
Test: manually
Change-Id: Ie9b5663af2e23ca0f7cfbc46e9a320d98401ca4e