Fix test apexes for size alignment
APEXes should be 4096-byte size-aligned. Well-formed APEXes are okay
because signapk handles this. But many test APEXes are hand-crafted with
genrule and they are not properly sized. This is normally okay for
tests that don't rely on size requirements. But with
go/mount-apex-before-data, this requirement is important becase APEX
files need to be mapped to dm-linear block devices.

Added a new tool to fix the test apexes for size requirements, which
just amends EOCD comments.

Bug: 423519370
Test: ApexTestCases
Change-Id: I072c8c5c826828930f53860e255ed8f1d18e5ead