apexd: loosen no-overlap check on install
Each staged-install session starts with "SubmitStagedSession" and ends
with either of "MarkStagedSessionReady" or "AbortStagedSession".
No-overlap check is to avoid new installs before the submission ends
with either success (MarkStagedSessionReady) or failure
(AbortStagedSession). But for some reason, we found there can be a
session that's submitted but not finished. This in effect disables
further installs.

Until we have a proper fix for this, let's loosen the check.

Bug: 424874098
Bug: 425478146
Test: ApexTestCases
Change-Id: I47f92390717eeb9d87a5a0e173e6644364a76021