apexd: clean up sessions after OTA
apexd marks old sessions as failure. But until now, only STAGED sessions
are covered for this check. We'd better checks *all* sessions if they
are created by old system (before OTA) and remove (or mark as failed)
sessions created by the old system.

In a recent bugreport, there was a three-year old session which is
in "VERIFIED" state for some reason. Note that VERIFIED sessions were
not handled by Build-fingerprint check.

In this commit, the check covers *all* sessions.

Bug: 425478146
Test: ApexTestCases
Change-Id: I3114e9a1202aec6ac2dd1318cbe710669b3d6710