apexd: Determine allocation alignment of pinned files (1/n)
Pinned files are created with allocation alignment. For example, f2fs
creates a pinned file with 2 MiB alignment. The alignment can be even
bigger to be 1 GiB on a certain device. ApexImageManager should adapt to
the alignment to avoid wasting too much space.

In this change, ApexImageManager determines the allocation alignment
of the /data filesystem and saves it in its metadata.

More changes will follow to extend ApexImageManager to use different
strategy based on the allocation alignment.

Bug: 402256229
Flag: com.android.apex.flags.mount_before_data
Test: atest --test-filter=*PinApexFiles
Change-Id: I56e1bb9f50f22b164b00854af27e0a4a710a4815