apexd: Save changed_active_apexes in sysprop
The list was to fill the ApexInfo.activeApexChanged field. But since it
was stored in a global variable, its correctness relies on the
implementation details:
- getActivePackages() should be served by the right process that
  activats the ALL apexes. A new instance of apexservice (as a lazy
  service) doesn't know about it.
- Since PM calls getActivePackages() before calling markBootCompleted(),
  it's the right one (for now, apexd that activates ALL apexes lives
  until markBootCompleted() is called by PM.)

This will break with mount_before_data. It's apexd-bootstrap that
activates ALL apexes (hence, knows what are "changed active apexes").
When PM calls getActivePackages(), it's a different process that serves
apexservice.

In this change, "changed_active_apexes" is stored in a sysprop and then
apexservice relies on it when answering getActivePackages(), instead of
a global variable.

Bug: 386290694
Flag: com.android.apex.flags.mount_before_data
Test: ApexTestCases
Change-Id: I30252e7972bb1815b3e136c3273280623831c170