apexd: UnstagePackages() for mount_before_data
Even though UnstagePackages() is not for production, it's used by many
tests to uninstall test apexes. The feature mount_before_data should
support it to keep the tests happy.

When mount_before_data is enabled, it removes the image names from the
ACTIVE list so that those images are not activated after reboot and
cleaned up as part of boot-completion cleanup.

Bug: 422941879
Flag: com.android.apex.flags.mount_before_data
Test: ApexTestCases
Change-Id: Ifae0a60dfc7cef5dce2d045233e80736ebe00a7f