Remove redundant --also-include-staged-apexes from --unmount-all command
and remove the path check from MountedApexDatabase::PopulateFromMounts().

PopulateFromMounts() doesn't need to check the path prefix for data
APEXes. The reason why the check was added was to distinguish two loop
devices for a single APEX mount when apexd supported zero hashtree. (To
support zero hashtree, the hashtree was generated on demand and two loop
devices were created for a single APEX mount. Path prefixes were used to
distinguish the hashtree loop device from the payload loop device.)

Since zero hashtree is no longer supported, the check is unnecessary.

`apexd --unmount-all` is to unmount ALL active mounts. The parameter
(`--also-include-staged-apexes`) was added as a counterpart of the
sub-mode(`--also-include-staged-apexes`) of the `--otachroot-bootstrap`
command. This was added to pass the path prefix check of
MountedApexDatabase::PopulateFromMounts().

This change removes both the prefix check and the redandant
parameter(--also-include-staged-apexes).

Bug: 365000206
Test: ApexTestCases
Change-Id: I4a4d195e99788242395a1587e4d43de4ff17f00b