Fix /data/apex/active references
In VendorApexHostTests, /data/apex/active path prefix was used to check
if the apex is from the data partition. It will be changed with
mount_before_data, hence the check is changed to use
ApplicationInfo.FLAG_UPDATED_SYSTEM_APP, which is more robust than
checking the path.

In ApexRollbackTests, /data/apex/active path is used to see if
BootCompletionCleanup removes inactive APEX files on. This is okay
because even with mount_before_data, the cleanup should remove those
files and the test can verify the expected behavior. Added more comments
to explain what's happening and removed unnecessary assertion for
readability.

Note that not all tests in ApexRollbackTests are passing with
mount_before_data enabled. It's tracked by b/431895953.

Bug: 430161238
Test: VendorApexHostTests
Test: ApexRollbackTests#testBootCompletedCleanupHappensEvenWhenThereIsCrashingProcess
Flag: com.android.apex.flags.mount_before_data
Change-Id: Iae9a5e4716b0d1ae836f7ee3969f5d9918384375