apexd: reduce I/O layers for APEX payload
Previously, a downloaded apex file is mapped as a dm-linear device on
top of the data partition. Another dm-linear device is created on top
of the first dm-linear to handle the offset and length for the
apex payload.

With the change, the second dm-linear is created directly on top of
the data partition. This reduces the overall I/O layers needed and
improves read bandwidth by 5% to 20% (FIO benchmark), depending on
the situation (apex size, number of job reading, etc).
go/apex-payload-perf-comparison for the detailed results.

Bug: 401446425
Test: flash the device
Change-Id: I03439a5b3059588ad96f82adb7a215fae143979c