apexd: hook commands for submitStagedSession
This is to make it easier to test the interaction between apexd and
PackageInstallerService. Root-privileged test driver can set sysprop
apexd.test_hook.submit_staged_session before committing a staged install
session. apexd runs the hook commands only if the CTS SHIM APEX is
installed on a debuggable device.

For now the available commands are `sleep_ms <ms>` and `error
<message>`.

Bug: 425478146
Test: StagedInstallInternalTest
Change-Id: I3e1d30bf0499bda46e3077b0e89ed87f704ff9b5