Reset backup on MarkStagedSessionSuccessful
Staged install creates a backup of the current active apexes. When the
install session is marked as successful, the backup should be cleared.

This patch adds the same behavior when mount_before_data is enabled.

In fact, the backup doesn't do any harm even if it's not cleared because
it's just no longer used and overwritten when a new install starts.
This patch is just to keep the old behavior with the new feature
enabled.

Bug: 408119691
Test: ApexTestCases
Flag: com.android.apex.flags.mount_before_data
Change-Id: Ifc2074b4189b8583b0f564f758277411fd51d2dc