Fix non-staged installation
There's a known issue that device-mapper devices are not deleted
properly when unmounting the APEX. With mount-before-data, repeated
non-staged installations fail because of the name conflicts of
DM-devices.

For example, let's say com.android.foo is installed multiple times.

- For the first installation, the apex is copied to "com.android.foo_0".
  Note that the name is used in many places (image manager, DM devices,
  etc)
- For the second installation, the incoming apex is copied to
  "com.android.foo_1". But unmounting "com.android.foo_0" leaves DM
  device undeleted while the image itself is deleted from the storage.
- Now the third installation fails because apexd thinks _0 slot is empty
  and use it for the incoming apex. But "com.android.foo_0" DM device is
  still live and used by other processes. (See b/467824824 for details)

Without mount-before-data, it was okay because DM devices are always
named with increasing suffix (_0, _1, _2, etc).

In this change, the APEX image manager is modified to keep the deleted
names so that the incoming APEXes don't reuse the deleted names. The
name conflicts only happen in the same boot. Reboot will clean up the
list of deleted images.

Bug: 467413483
Test: repeat "adb install --non-staged"
Flag: com.android.apex.flags.mount_before_data
Change-Id: I74171e7ea48cdc9c7b34aa05117285ac2254c805