Disable v1 signatures in apexer tests
Pass --disable-v1 to signapk.jar in apexer tests to match the
build system.

Use zip2zip instead of zip to gracefully continue when the input
zip has now META-INF/MANIFEST.MF file.  The V2+ signatures are
stored in a binary blob located before the zip central directory
that is outside any of the files, so they are always stripped
when passing a zip through zip2zip.

Bug: 325130426
Test: apexer_test
Change-Id: I0005580f0c7216e696592c7e37b109bdcc26c806