Wait 1s until the DM device is deleted
This fixes the test flakiness. The test checks if non-staged
installation removes the old apex. With mount_before_data, the removal
means the deletion of DM device. The device removal takes some time
until the uevent is handled by ueventd. This fixes the flakiness in the
test that checks if the old APEX path is removed.

Adding 1s timeout when deleting the device doesn't harm in normal cases
because it only happens in an error case or non-staged installation.

Bug: n/a
Flag: com.android.apex.flags.mount_before_data
Test: ApexTestCases:NonStagedInstall_SucceedAgainstData
Change-Id: Ifad5cf3c4edab5a40caed492a03fa7007fc2ef3f